Update naive version to get correct result.

+++ 01_naive_atomic.cu
@@ -1,9 +1,9 @@
 #include <cstdio>
 
-__global__ void reduceNaive(const float* in, float* out, int n) {
+__global__ void reduceNaive(const float* in, double* out, int n) {
     int i = blockIdx.x * blockDim.x + threadIdx.x;
     if (i < n)
-        atomicAdd(out, in[i]);   // every thread hammers the same address
+        atomicAdd(out, (double)in[i]);   // every thread hammers the same address
 }
 
 int main() {
@@ -16,15 +16,15 @@ int main() {
 
     // 1. Allocate host (CPU) memory and fill it
     float *h_A = (float*)malloc(bytes);
-    float h_out = 0.0f;
+    double h_out = 0.0;
     for (int i = 0; i < n; i++) { h_A[i] = 1.0f;}
 
     // 2. Allocate device (GPU) memory
     float *d_A;
-    float *d_out;
+    double *d_out;
     cudaMalloc(&d_A, bytes);
-    cudaMalloc(&d_out, sizeof(float));
-    cudaMemset(d_out, 0, sizeof(float));
+    cudaMalloc(&d_out, sizeof(double));
+    cudaMemset(d_out, 0, sizeof(double));
 
     // 3. Copy inputs from host to device
     cudaMemcpy(d_A, h_A, bytes, cudaMemcpyHostToDevice);
@@ -35,7 +35,7 @@ int main() {
     reduceNaive<<<blocks, threadsPerBlock>>>(d_A, d_out, n);
     cudaDeviceSynchronize();
     
-    cudaMemset(d_out, 0, sizeof(float));
+    cudaMemset(d_out, 0, sizeof(double));
     cudaEventRecord(start);
     reduceNaive<<<blocks, threadsPerBlock>>>(d_A, d_out, n);
     cudaEventRecord(stop);
@@ -46,13 +46,14 @@ int main() {
 
 
     // 5. Copy result back from device to host
-    cudaMemcpy(&h_out, d_out, sizeof(float), cudaMemcpyDeviceToHost);
+    cudaMemcpy(&h_out, d_out, sizeof(double), cudaMemcpyDeviceToHost);
 
     // 6. Verify (every element should be 1.0)
     bool ok = true;
     if (h_out != 32000000.0f) { ok = false; }
     printf("Result: %f\n", h_out);
     printf("%s\n", ok ? "PASS: all elements correct" : "FAIL");
+
 
     // 7. Free memory
     cudaFree(d_A); cudaFree(d_out);
